backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.8.1
celery==5.4.0
redis==5.2.1
python-multipart==0.0.20
slowapi==0.1.9
alembic==1.14.1
email-validator==2.2.0
reportlab==4.2.5
httpx==0.28.1
websockets==13.1
google-generativeai==0.8.4
